+++ src/main/java/org/scion/demo/ScionPathExample.java
@@ -21,6 +21,7 @@ import org.scion.jpan.*;
 
 public class ScionPathExample {
   public static void main(String[] args) throws IOException {
+    String message = args.length == 0 ? "Hello there!" : args[0];
     // send to https://scionpacketinspector.netsec.ethz.ch/
     long dstIsdAs = ScionUtil.parseIA("64-2:0:9");
     InetAddress dstIP = InetAddress.getByName("129.132.175.104");
@@ -28,7 +29,7 @@ public class ScionPathExample {
 
     try (ScionDatagramChannel channel = ScionDatagramChannel.open()) {
       channel.connect(path);
-      ByteBuffer sendBuf = ByteBuffer.wrap("Hello there!".getBytes());
+      ByteBuffer sendBuf = ByteBuffer.wrap(message.getBytes());
       channel.write(sendBuf);
       PathMetadata meta = channel.getConnectionPath().getMetadata();
       System.out.println("Sent via path: " + ScionUtil.toStringPath(meta));